requirements.txt
asyncio==3.4.3
certifi==2024.7.4
charset-normalizer==3.3.2
idna==3.7
pillow==11.2.1
pyTelegramBotAPI==4.21.0
requests==2.32.4
urllib3==2.2.2
